Move related components to subpackage
