+++ src/native-ads/NativeAdsManager.ts
